+++ packages/wrapper/web/src/localeStrategy/localeStrategyProvider.js
@@ -1,10 +1,12 @@
-export default ({ locale = 'en' } = { locale: 'en' }) => (innerProps = {}) => {
+export default ({ locale = 'en' } = { locale: 'en' }) => async (innerProps = {}) => {
   const mergedLocale = innerProps.locale || locale;
   if (mergedLocale === 'en') {
-    return { locale: mergedLocale, ...innerProps };
+    return innerProps;
   }
   try {
-    const localeResource = require(`wix-rich-content-common/dist/statics/locale/messages_${mergedLocale}.json`); // eslint-disable-line max-len
+    const localeResource = await import(
+      `wix-rich-content-common/dist/statics/locale/messages_${mergedLocale}.json`
+    ).then(res => res.default);
     return { locale: mergedLocale, localeResource, ...innerProps };
   } catch (err) {
     throw new Error(`error while loading locale ${locale}`, err);